Build the entire thing with rollup
