+++ .gitignore
@@ -1,4 +1,5 @@
 /coverage
+/dist
 /node_modules
 /.nyc_output
 /package-lock.json
