+++ package.json
@@ -15,33 +15,37 @@
   "keywords": [
     "dom"
   ],
-  "main": "./pakertaja.min.js",
+  "main": "dist/pakertaja.js",
+  "module": "dist/pakertaja.mjs",
+  "browser": "dist/pakertaja.iife.js",
+  "types": "index.d.ts",
   "scripts": {
-    "build": "babel pakertaja.js -o pakertaja.min.js",
+    "build": "rollup -c",
+    "watch": "rollup -c -w",
     "pretest": "yarn run build",
     "prepare": "yarn run build",
     "test": "nyc --reporter=html --reporter=text mocha",
     "coverage": "nyc report --reporter=text-lcov | coveralls"
   },
   "devDependencies": {
-    "@babel/cli": "^7.8.4",
     "@babel/core": "^7.9.6",
     "@babel/preset-env": "^7.9.6",
+    "@rollup/plugin-babel": "^5.0.2",
     "babel-preset-minify": "^0.5.1",
     "coveralls": "^3.1.0",
     "mocha": "^7.1.2",
     "nodom": "^2.4.0",
     "nyc": "^15.0.1",
+    "rollup": "^2.10.5",
     "should": "^13.2.3",
     "yarn": "^1.22.4"
   },
-  "types": "./index.d.ts",
   "files": [
     "LICENSE",
     "README.md",
     "index.d.ts",
-    "pakertaja.js",
-    "pakertaja.min.js",
-    "pakertaja.min.js.map"
+    "dist/pakertaja.js",
+    "dist/pakertaja.mjs",
+    "dist/pakertaja.iife.js"
   ]
 }
