+++ test.js
@@ -4,7 +4,7 @@ const nodom = require('nodom');
 global.document = new nodom.Document();
 
 describe('Pakertaja', () => {
-  const p = require('./pakertaja');
+  const p = require('./dist/pakertaja');
 
   it('should be able to create DOM elements', () => {
     const elements = {
